refactor(cloudevents-server): isolate Jenkins routing policy
