+++ cloudevents-server/pkg/events/handler/jenkins_routing_test.go
@@ -16,9 +16,9 @@ func TestResolveTopic(t *testing.T) {
 	}
 
 	tests := []struct {
-		name      string
-		eventType string
-		wantTopic string
+		name       string
+		eventType  string
+		wantTopic  string
 		wantIgnore bool
 	}{
 		{
